navbar/demo/requirements.txt
gradio_navbar